1_LangChain/2_Prompts/requirements.txt
langchain
langchain-groq
faiss-cpu
tiktoken
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
pymupdf
python-docx
docx2txt
unstructured[docx]
openpyxl
unstructured[xlsx]
jq
matplotlib
langchain_experimental
rank_bm25
wikipedia
hf_xet
InstructorEmbedding
##Embeddings
langchain-huggingface
sentence-transformers
streamlit